+++ lib/ui/onboarding/OnboardingLoginPhoneConfirmPanel.dart
@@ -16,7 +16,6 @@
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/model/UserData.dart';
 import 'package:illinois/service/Auth.dart';
 import 'package:illinois/service/Onboarding.dart';
 import 'package:illinois/service/Localization.dart';
@@ -43,7 +42,7 @@ class OnboardingLoginPhoneConfirmPanel extends StatefulWidget with OnboardingPan
 
   @override
   bool get onboardingCanDisplay {
-    return !User().roles.contains(UserRole.student) && !User().roles.contains(UserRole.employee) && (onboardingContext != null) && onboardingContext['shouldVerifyPhone'] == true;
+    return !User().isStudentOrEmployee && (onboardingContext != null) && onboardingContext['shouldVerifyPhone'] == true;
   }
 }