+++ lib/ui/onboarding/OnboardingLoginPhoneVerifyPanel.dart
@@ -17,7 +17,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/model/UserData.dart';
 import 'package:illinois/service/Auth.dart';
 import 'package:illinois/service/Onboarding.dart';
 import 'package:illinois/service/Localization.dart';
@@ -42,7 +41,7 @@ class OnboardingLoginPhoneVerifyPanel extends StatefulWidget with OnboardingPane
 
   @override
   bool get onboardingCanDisplay {
-    return !User().roles.contains(UserRole.student) && !User().roles.contains(UserRole.employee) && (onboardingContext != null) && onboardingContext['shouldVerifyPhone'] == true;
+    return !User().isStudentOrEmployee && (onboardingContext != null) && onboardingContext['shouldVerifyPhone'] == true;
   }
 }