+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Empty Text for Covid19TestLocations panel & show wait time status (long/medium/short) [#278](https://github.com/rokwire/safer-illinois-app/issues/278)
 - Remove private params from the url. Additional fix [#110](https://github.com/rokwire/safer-illinois-app/issues/110).
 - Add external link icon for SettingsHomePanel buttons [241](https://github.com/rokwire/safer-illinois-app/issues/241).
+- Store RSA private key separately for organization and environment.
 
 ## [2.7.6] - 2020-10-13
 ### Changed